Add Design Creator/Item
## Description

Sometimes it's hard to keep track of creators
and items we want to follow or clothes and
patterns we want to use.

A good place to track these information would be
in the companion application.

## Implementation

The main types added by this commit are:

- `Design`: simple model with title, code,
   and description.
- `DesignRowView` and `DesignRowViewModel`: view
   and view model used tp populate the table
- `DesignFormView` and `DesignFormViewModel`: view
   and view model used to add a new design to
   the list.

This commit also touches the caching layer, adding
designs as an optional array.